refactor: replace `ModuleData` with `ModuleSettings` in `RoomParameters`
Adds a new `ModuleSettings` struct that contains the settings of the
signaling modules. This struct is equivalent to the existing `ModuleData`
that we were using in the `RoomParameters` before. Instead of the
`SignalingModuleFrontendData` trait it is making use of the new
`SignalingModuleSetting` trait for accessing module specific settings.
This allows us to clearly differentiate between the data meant for the
backend and the one meant for the frontend.

Closes #144